Misuv 4863 annuity/royalty payments (#1600)
* MISUV-4796 CL279 - Add two fields (Annuity/Royalty payments) and display pence values in Tax calculation page

* MISUV-4796 Support decimal values for totalAnnuityPaymentsTaxCharged

+++ app/models/liabilitycalculation/taxcalculation/TaxCalculation.scala
@@ -27,7 +27,7 @@ case class TaxCalculation(
                            totalIncomeTaxAndNicsDue: BigDecimal,
                            totalTaxDeducted: Option[BigDecimal] = None,
                            totalIncomeTaxAndNicsAndCgt:Option[BigDecimal] = None,
-                           totalAnnuityPaymentsTaxCharged: Option[Int] = None,
+                           totalAnnuityPaymentsTaxCharged: Option[BigDecimal] = None,
                            totalRoyaltyPaymentsTaxCharged: Option[BigDecimal] = None
                          )
